+++ Js2IL.Tests/JavaScript/ControlFlow_If_Truthiness.js
@@ -1,0 +1,13 @@
+let url = "";
+if (url) {
+  console.log("truthy");
+} else {
+  console.log("falsy");
+}
+
+url = "http://example.com";
+if (url) {
+  console.log("truthy");
+} else {
+  console.log("falsy");
+}